Cross industry processes

+++ src/scripts/processes.ts
@@ -12,38 +12,47 @@ async function main() {
   if (!connection.isConnected) {
     throw new Error('connection is not established');
   }
+  let industries: any = [];
   let root = null;
   let data: any = [];
   let resultDeleted = null;
 
+  const industryName = process.argv[2];
+
   await getManager().transaction(async transactionalEntityManager => {
-    let industries: any = await parseCsv('industries.csv', rows =>
-      rows.map((row: any) => ({
-        ...row,
-        id: +row.id,
-      }))
-    );
-    for (let industry of industries) {
-      const found = await transactionalEntityManager.findOne(Industry, {
-        where: { name: industry.name },
+    if (industryName) {
+      industries = await transactionalEntityManager.find(Industry, {
+        where: { name: industryName },
       });
-      if (!found) {
-        await transactionalEntityManager
-          .createQueryBuilder()
-          .insert()
-          .into('industries', ['id', 'name'])
-          .values([industry])
-          .execute();
+    } else {
+      industries = await parseCsv('industries.csv', rows =>
+        rows.map((row: any) => ({
+          ...row,
+          id: +row.id,
+        }))
+      );
+      for (let industry of industries) {
+        const found = await transactionalEntityManager.findOne(Industry, {
+          where: { name: industry.name },
+        });
+        if (!found) {
+          await transactionalEntityManager
+            .createQueryBuilder()
+            .insert()
+            .into('industries', ['id', 'name'])
+            .values([industry])
+            .execute();
+        }
       }
+      industries = await transactionalEntityManager.find(Industry);
     }
 
-    // get industries
-    industries = await transactionalEntityManager.find(Industry);
-
-    // clear all processes
-    resultDeleted = await transactionalEntityManager.delete(Process, {});
-    if (resultDeleted) {
-      console.log(`Process: Number of deleted records ${resultDeleted.affected}`);
+    if (!industryName) {
+      // clear all processes
+      resultDeleted = await transactionalEntityManager.delete(Process, {});
+      if (resultDeleted) {
+        console.log(`Process: Number of deleted records ${resultDeleted.affected}`);
+      }
     }
 
     // save tree processes for each industry
